feat(HomeScreen.tsx): add support for dynamic image carousel on the home screen
The changes in this commit add support for a dynamic image carousel on the home screen. Four new images (photo2.png, photo3.png, photo4.jpg) were added to the carousel. The carousel allows the user to click on the image to cycle through the images. The animation for transitioning between images was also added to provide a smooth visual experience.