feat(robots.txt): add robots.txt file to allow all user agents and disallow nothing
The robots.txt file was added to the project. This file is used to communicate with web crawlers and specify which parts of the website should not be crawled or indexed. In this case, the file allows all user agents and disallows nothing, meaning that all parts of the website can be crawled and indexed by search engines.