feat(navbar): add logo and language selector to mobile navbar
The changes in this commit add a logo and a language selector to the mobile navbar component.

The logo is imported from the `logo.svg` file located in the `@/public` directory. It is displayed using the `next/image` component with a width of 127 pixels.

The language selector component (`Language`) is imported from the `@/components/[locale]/desktop/navigation/navbar` directory. It is added to the navbar component using a `div` element with the `flex justify-between p-6` class.